deploy/requirements.txt
py-clob-client>=0.34.0
python-dotenv>=1.0.0
requests>=2.31.0
matplotlib>=3.7.0
